Add guardrails to ensure proper typing and listing of certain types of DB columns.
