+++ docs/Troubleshooting.md
@@ -23,7 +23,7 @@ Make sure that an OIDC App has been created, and that the `Client ID` parameter 
 
 ### Invalid Redirect Uri
 
-The 'Client' application should specify the correct `Redirect URL` parameter for entry into VoidAuth. This is commonly located in the OIDC documentation of the application or on it's OIDC Setup page if it has one. You may also be able to find an example for the application on the [OIDC App Guides](OIDC-Guides.md) page.
+The 'Client' application should specify the correct `Redirect URL` parameter for entry into VoidAuth. This is commonly located in the OIDC documentation of the application or on its OIDC Setup page if it has one. You may also be able to find an example for the application on the [OIDC App Guides](OIDC-Guides.md) page.
 
 ### The Page Cannot Be Found
