+++ frontend/src/app/pages/admin/domains/domain/domain.component.ts
@@ -10,7 +10,8 @@ import type { ProxyAuthUpsert } from '@shared/api-request/admin/ProxyAuthUpsert'
 import { CommonModule } from '@angular/common'
 import { MaterialModule } from '../../../../material-module'
 import { ValidationErrorPipe } from '../../../../pipes/ValidationErrorPipe'
-import { isValidWildcardDomain, type Nullable } from '@shared/utils'
+import { type Nullable } from '@shared/utils'
+import { isValidWildcardDomain } from '@shared/url'
 import type { ProxyAuthResponse } from '@shared/api-response/admin/ProxyAuthResponse'
 import { MatDialog } from '@angular/material/dialog'
 import { ConfirmComponent } from '../../../../dialogs/confirm/confirm.component'
