+++ frontend/src/app/pages/admin/domains/domains.component.ts
@@ -9,7 +9,7 @@ import { SpinnerService } from '../../../services/spinner.service'
 import type { TableColumn } from '../clients/clients.component'
 import { RouterLink } from '@angular/router'
 import { MaterialModule } from '../../../material-module'
-import { sortWildcardDomains } from '@shared/utils'
+import { sortWildcardDomains } from '@shared/url'
 import { MatDialog } from '@angular/material/dialog'
 import { ConfirmComponent } from '../../../dialogs/confirm/confirm.component'
 import { TranslatePipe } from '@ngx-translate/core'
