+++ server/db/tableMaintenance.ts
@@ -2,7 +2,7 @@ import type { Key } from '@shared/db/Key'
 import { db } from './db'
 import type { Consent } from '@shared/db/Consent'
 import type { EmailVerification } from '@shared/db/EmailVerification'
-import { TABLES } from '@shared/constants'
+import { TABLES } from '@shared/db'
 import type { Invitation } from '@shared/db/Invitation'
 import type { OIDCPayload } from '@shared/db/OIDCPayload'
 import type { PasskeyRegistration, PasskeyAuthentication } from '@shared/db/Passkey'
