+++ server/db/totp.ts
@@ -1,11 +1,12 @@
 import * as OTPAuth from 'otpauth'
 import { db } from './db'
 import type { TOTP } from '@shared/db/TOTP'
-import { TABLES, TTLs } from '@shared/constants'
+import { TTLs } from '@shared/constants'
 import appConfig from '../util/config'
 import { randomUUID } from 'crypto'
 import { createExpiration, decryptString, encryptString } from './util'
 import type { RegisterTotpResponse } from '@shared/api-response/RegisterTotpResponse'
+import { TABLES } from '@shared/db'
 
 function decryptTOTP(totp: TOTP | undefined): TOTP | null | undefined {
   if (!totp) {
