+++ server/oidc/provider.ts
@@ -2,7 +2,7 @@ import Provider, { type Configuration } from 'oidc-provider'
 import { findAccount, getUserById, userRequiresMfa } from '../db/user'
 import appConfig, { basePath, getSessionDomain, sessionDomainReaches } from '../util/config'
 import { KnexAdapter } from './adapter'
-import { ADMIN_GROUP, REDIRECT_PATHS, TABLES, TTLs } from '@shared/constants'
+import { ADMIN_GROUP, REDIRECT_PATHS, TTLs } from '@shared/constants'
 import { errors } from 'oidc-provider'
 import { getCookieKeys, getJWKs, makeKeysValid } from '../db/key'
 import Keygrip from 'keygrip'
@@ -10,7 +10,7 @@ import { interactionPolicy } from 'oidc-provider'
 import { isExpired, isUnapproved, isUnverifiedEmail, loginFactors } from '@shared/user'
 import { isMatch } from 'matcher'
 import assert from 'assert'
-import { wildcardRedirect } from '@shared/utils'
+import { wildcardRedirect } from '@shared/url'
 import type { IncomingMessage, ServerResponse } from 'http'
 import { getProxyAuthWithCache } from '../db/proxyAuth'
 import { RESPONSE_TYPES } from '@shared/api-request/admin/ClientUpsert'
@@ -24,6 +24,7 @@ import { db } from '../db/db'
 import { mergeKeys } from '../db/util'
 import type { User } from '@shared/db/User'
 import { PayloadTypes } from '@shared/db/OIDCPayload'
+import { TABLES } from '@shared/db'
 
 // Extend 'oidc-provider' where needed
 declare module 'oidc-provider' {
