+++ server/routes/admin.ts
@@ -7,7 +7,7 @@ import { randomBytes, randomUUID } from 'crypto'
 import { getClient, getClients } from '../db/client'
 import type { UserGroup, Group, InvitationGroup, ProxyAuthGroup } from '@shared/db/Group'
 import { groupUpsertValidator } from '@shared/api-request/admin/GroupUpsert'
-import { ADMIN_GROUP, TABLES, TTLs } from '@shared/constants'
+import { ADMIN_GROUP, TTLs } from '@shared/constants'
 import { userUpdateValidator } from '@shared/api-request/admin/UserUpdate'
 import { endSessions, getUserById, getUsers } from '../db/user'
 import { createExpiration, mergeKeys } from '../db/util'
@@ -18,7 +18,7 @@ import { invitationUpsertValidator } from '@shared/api-request/admin/InvitationU
 import { sendApproved, sendInvitation, sendPasswordReset, sendTestNotification, SMTP_VERIFIED } from '../util/email'
 import type { GroupUsers } from '@shared/api-response/admin/GroupUsers'
 import type { ProxyAuth } from '@shared/db/ProxyAuth'
-import { urlFromWildcardHref } from '@shared/utils'
+import { urlFromWildcardHref } from '@shared/url'
 import type { ProxyAuthResponse } from '@shared/api-response/admin/ProxyAuthResponse'
 import { proxyAuthUpsertValidator } from '@shared/api-request/admin/ProxyAuthUpsert'
 import { getProxyAuth, getProxyAuths } from '../db/proxyAuth'
@@ -37,6 +37,7 @@ import zod from 'zod'
 import { checkAdmin, checkPrivileged } from '../util/authMiddleware'
 import type { AdminConfig } from '@shared/api-response/admin/AdminConfig'
 import type { IncomingMessage } from 'http'
+import { TABLES } from '@shared/db'
 
 export const adminRouter = Router()
