+++ server/routes/public.ts
@@ -5,7 +5,7 @@ import { sendPasswordReset, SMTP_VERIFIED } from '../util/email'
 import type { PasswordReset } from '@shared/db/PasswordReset'
 import { endSessions, getUserById, getUserByInput } from '../db/user'
 import { db } from '../db/db'
-import { TABLES } from '@shared/constants'
+import { TABLES } from '@shared/db'
 import type { SendPasswordResetResponse } from '@shared/api-response/SendPasswordResetResponse'
 import { resetPasswordValidator } from '@shared/api-request/ResetPassword'
 import type { User } from '@shared/db/User'
