+++ server/routes/user.ts
@@ -10,7 +10,7 @@ import type { User } from '@shared/db/User'
 import { checkPasswordHash } from '../db/user'
 import { deleteUserPasskey, deleteUserPasskeys, getUserPasskeys, getUserPasskeysResponse } from '../db/passkey'
 import type { OIDCPayload } from '@shared/db/OIDCPayload'
-import { TABLES } from '@shared/constants'
+import { TABLES } from '@shared/db'
 import type { TOTP } from '@shared/db/TOTP'
 import { argon2 } from '../util/argon2id'
 import { zodValidate } from '../util/zodValidate'
