+++ server/util/email.ts
@@ -4,7 +4,7 @@ import nodemailer from 'nodemailer'
 import pug from 'pug'
 import appConfig from './config'
 import type SMTPTransport from 'nodemailer/lib/smtp-transport'
-import { ADMIN_GROUP, REDIRECT_PATHS, TABLES } from '@shared/constants'
+import { ADMIN_GROUP, REDIRECT_PATHS } from '@shared/constants'
 import type { Invitation } from '@shared/db/Invitation'
 import type { PasswordReset } from '@shared/db/PasswordReset'
 import { PRIMARY_CONTRAST_COLOR, PRIMARY_COLOR } from './theme'
@@ -16,6 +16,7 @@ import type { UserWithoutPassword } from '@shared/api-response/UserDetails'
 import ejs from 'ejs'
 import { logger } from './logger'
 import { getPasswordResetURL } from '../db/passwordReset'
+import { TABLES } from '@shared/db'
 
 export let SMTP_VERIFIED = false
 const DEFAULT_EMAIL_TEMPLATE_DIR = './default_email_templates'
