+++ shared/constants.ts
@@ -1,5 +1,3 @@
-import type { ValueOf } from './utils'
-
 export const ADMIN_GROUP = 'auth_admins' as const
 export const ADMIN_USER = 'auth_admin' as const
 
@@ -60,51 +58,3 @@ export const TTLs = {
   DAY: 1 * DAY,
   YEAR: 1 * YEAR,
 } as const
-
-export const TABLES = {
-  KEY: 'key',
-  FLAG: 'flag',
-  USER: 'user',
-  GROUP: 'group',
-  USER_GROUP: 'user_group',
-  CONSENT: 'consent',
-  EMAIL_VERIFICATION: 'email_verification',
-  PASSWORD_RESET: 'password_reset',
-  INVITATION: 'invitation',
-  INVITATION_GROUP: 'invitation_group',
-  PROXY_AUTH: 'proxy_auth',
-  PROXY_AUTH_GROUP: 'proxy_auth_group',
-  OIDC_PAYLOADS: 'oidc_payloads',
-  PASSKEY: 'passkey',
-  PASSKEY_REGISTRATION: 'passkey_registration',
-  PASSKEY_AUTHENTICATION: 'passkey_authentication',
-  EMAIL_LOG: 'email_log',
-  OIDC_GROUP: 'oidc_group',
-  TOTP: 'totp',
-} as const
-
-export const TABLES_ORDER: ValueOf<typeof TABLES>[] = [
-  TABLES.KEY,
-  TABLES.FLAG,
-  TABLES.USER,
-  TABLES.GROUP,
-  TABLES.USER_GROUP,
-  TABLES.CONSENT,
-  TABLES.EMAIL_VERIFICATION,
-  TABLES.PASSWORD_RESET,
-  TABLES.INVITATION,
-  TABLES.INVITATION_GROUP,
-  TABLES.PROXY_AUTH,
-  TABLES.PROXY_AUTH_GROUP,
-  TABLES.OIDC_PAYLOADS,
-  TABLES.PASSKEY,
-  TABLES.PASSKEY_REGISTRATION,
-  TABLES.PASSKEY_AUTHENTICATION,
-  TABLES.EMAIL_LOG,
-  TABLES.OIDC_GROUP,
-  TABLES.TOTP,
-] as const
-
-if (!Object.values(TABLES).every(t => TABLES_ORDER.includes(t))) {
-  throw new Error('Migration tables list is missing tables.')
-}
