+++ shared/db/EmailVerification.ts
@@ -1,3 +1,4 @@
+import type { DBColumnTypesCheck } from '@shared/db.js'
 import type { Audit } from './Audit.js'
 import type { User } from './User.js'
 
@@ -8,3 +9,5 @@ export type EmailVerification = Pick<Audit, 'createdAt'> & {
   challenge: string
   expiresAt: Date | number
 }
+
+const _typeCheck: DBColumnTypesCheck<EmailVerification> = true
