+++ shared/db/Flag.ts
@@ -1,5 +1,9 @@
+import type { DBColumnTypesCheck } from '@shared/db'
+
 export type Flag = {
   name: string
   value: string | null
   createdAt: Date | number
 }
+
+const _typeCheck: DBColumnTypesCheck<Flag> = true
