+++ shared/db/PasswordReset.ts
@@ -1,5 +1,6 @@
-import type { Audit } from './Audit.js'
-import type { User } from './User.js'
+import type { DBColumnTypesCheck } from '@shared/db.js'
+import type { Audit } from './Audit'
+import type { User } from './User'
 
 export type PasswordReset = Pick<Audit, 'createdAt'> & {
   id: string
@@ -7,3 +8,5 @@ export type PasswordReset = Pick<Audit, 'createdAt'> & {
   challenge: string
   expiresAt: Date | number
 }
+
+const _typeCheck: DBColumnTypesCheck<PasswordReset> = true
