+++ shared/db/TOTP.ts
@@ -1,3 +1,4 @@
+import type { DBColumnTypesCheck } from '@shared/db'
 import type { Audit } from './Audit'
 
 export type TOTP = Pick<Audit, 'createdAt' | 'updatedAt'> & {
@@ -6,3 +7,5 @@ export type TOTP = Pick<Audit, 'createdAt' | 'updatedAt'> & {
   secret: string
   expiresAt: Date | number | null
 }
+
+const _typeCheck: DBColumnTypesCheck<TOTP> = true
